@@ -1,10 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace AwesomeBackend.BusinessLayer.Models
+namespace AwesomeBackend.Common.Models.Responses
 {
     public class NewRating
     {
